First Rough Draft of the Car Hierarchy
Each Part has been completed with the basic functionality they all share. More specific functionality will be added at a later stage.

+++ Car.h
@@ -75,7 +75,7 @@ class Car {
 			Sets the stored handling to the given value
 			@param
 		*/
-		virtual void setHandling( float ) = 0;
+		virtual void setHandling( int ) = 0;
 
 		/**
 			Returns the stored speed
@@ -86,7 +86,7 @@ class Car {
 			Sets the stored handling to the given value
 			@param
 		*/
-		virtual void setSpeed( float ) = 0;
+		virtual void setSpeed( int ) = 0;
 
 		/**
 			Returns the acceleration handling
@@ -97,7 +97,7 @@ class Car {
 			Sets the stored handling to the given value
 			@param
 		*/
-		virtual void setAcceleration( float ) = 0;
+		virtual void setAcceleration( int ) = 0;
 };
 
 #endif
